moved formatting functionality out of TextUI and into a new Formatter class
